feat(apparish): encrypt apparish file
The 'bm' command from apparish has been modified to 'apparishbm' in the
original source. 'bm' is now a function alias for 'apparishbm' plus
adding the file back into the chezmoi source encrypted. This is
probably the closest possible to symlink+encryption in chezmoi and will
do for now.

The apparixexpand file has been removed as it wasn't being used.

There was originally going to be several different apparixrc files
based on the architecture but at this point, it's easier to constrain
the usage based on os and make do.